ai_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
openai # <-- المكتبة الجديدة
python-multipart  # <-- هذا هو السطر الجديد الذي يحل المشكلة
